setting up parameters
