+++ config.toml
@@ -47,3 +47,11 @@ priority_fee_basis_points = 0
 liquidity_mantissa = 1
 liquidity_exponent = 20
 initial_price = 1.0
+
+[simulation_parameters]
+number_of_paths = 5
+
+    [simulation_parameters.sweep_parameters.volatility_basis_points]
+    start = 5
+    end = 100
+    steps = 10
